Switch minimap into a simple t_image                    [OK] Get player data                                         [OK] Update t_map attributes                                 [OK]   Use uint16_t instead of uint32_t for width and height [OK] Implement the minimap                                   [DOING]   Fill the image                                        [DOING]     Floor cells                                         [TODO]     Wall cells                                          [TODO]     Player                                              [TODO]       Position                                          [TODO]       Direction                                         [TODO]       Field of view                                     [TODO]

+++ private/t_map.h
@@ -6,7 +6,7 @@
 /*   By: jodufour <jodufour@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/14 01:14:23 by jodufour          #+#    #+#             */
-/*   Updated: 2023/05/14 23:44:12 by jodufour         ###   ########.fr       */
+/*   Updated: 2023/05/22 21:54:22 by jodufour         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,15 +15,14 @@
 
 # include "e_wall_texture.h"
 # include "t_image.h"
-# include <stdint.h>
 
 typedef struct s_map	t_map;
 
-struct	s_map
+struct s_map
 {
 	char		*cells;
-	uint32_t	width;
-	uint32_t	height;
+	uint16_t	width;
+	uint16_t	height;
 	uint32_t	floor_color;
 	uint32_t	ceiling_color;
 	t_image		wall_textures[WALL_TEXTURE_COUNT];
